libs/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared utilities for logging, configuration, and common helpers"
requires-python = ">=3.12"
dependencies = [
    "joblib>=1.5.3",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
